+++ cmd/template/app/flag.go
@@ -13,6 +13,7 @@ import (
 const (
 	flagAppName                    = "app-name"
 	flagCatalog                    = "catalog"
+	flagCatalogNamespace           = "catalog-namespace"
 	flagCluster                    = "cluster"
 	flagDefaultingEnabled          = "defaulting-enabled"
 	flagInCluster                  = "in-cluster"
@@ -20,6 +21,7 @@ const (
 	flagNamespace                  = "namespace"
 	flagNamespaceConfigAnnotations = "namespace-annotations"
 	flagNamespaceConfigLabels      = "namespace-labels"
+	flagOrganization               = "organization"
 	flagUserConfigMap              = "user-configmap"
 	flagUserSecret                 = "user-secret"
 	flagVersion                    = "version"
@@ -28,11 +30,13 @@ const (
 type flag struct {
 	AppName                        string
 	Catalog                        string
+	CatalogNamespace               string
 	Cluster                        string
 	DefaultingEnabled              bool
 	InCluster                      bool
 	Name                           string
 	Namespace                      string
+	Organization                   string
 	Version                        string
 	flagNamespaceConfigAnnotations []string
 	flagNamespaceConfigLabels      []string
@@ -43,9 +47,11 @@ type flag struct {
 func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.AppName, flagAppName, "", "Optionally set a different name for the App CR.")
 	cmd.Flags().StringVar(&f.Catalog, flagCatalog, "", "Catalog name where app is stored.")
+	cmd.Flags().StringVar(&f.CatalogNamespace, flagCatalogNamespace, "", "Catalog namespace where Catalog CR is stored if outside the default namespace.")
 	cmd.Flags().StringVar(&f.Name, flagName, "", "Name of the app in the Catalog.")
 	cmd.Flags().StringVar(&f.Namespace, flagNamespace, "", "Namespace where the app will be deployed.")
 	cmd.Flags().StringVar(&f.Cluster, flagCluster, "", "Name of the cluster the app will be deployed to.")
+	cmd.Flags().StringVar(&f.Organization, flagOrganization, "", "Workload cluster organization.")
 	cmd.Flags().BoolVar(&f.DefaultingEnabled, flagDefaultingEnabled, true, "Don't template fields that will be defaulted.")
 	cmd.Flags().BoolVar(&f.InCluster, flagInCluster, false, fmt.Sprintf("Deploy the app in the current management cluster rather than in a workload cluster. If this is set, --%s will be ignored.", flagCluster))
 	cmd.Flags().StringVar(&f.flagUserConfigMap, flagUserConfigMap, "", "Path to the user values configmap YAML file.")
@@ -53,6 +59,8 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.Version, flagVersion, "", "App version to be installed.")
 	cmd.Flags().StringSliceVar(&f.flagNamespaceConfigAnnotations, flagNamespaceConfigAnnotations, nil, "Namespace configuration annotations in form key=value.")
 	cmd.Flags().StringSliceVar(&f.flagNamespaceConfigLabels, flagNamespaceConfigLabels, nil, "Namespace configuration labels in form key=value.")
+
+	_ = cmd.Flags().MarkHidden(flagOrganization)
 }
 
 func (f *flag) Validate() error {