+++ cmd/template/app/runner.go
@@ -2,13 +2,14 @@ package app
 
 import (
 	"context"
+	"fmt"
 	"io"
 	"os"
+	"strings"
 	"text/template"
 
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 	"sigs.k8s.io/yaml"
 
@@ -21,8 +22,8 @@ import (
 type runner struct {
 	flag   *flag
 	logger micrologger.Logger
-	stdout io.Writer
 	stderr io.Writer
+	stdout io.Writer
 }
 
 func (r *runner) Run(cmd *cobra.Command, args []string) error {
@@ -51,14 +52,20 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		appName = r.flag.Name
 	}
 
+	// Since organization may be provided in a mixed-cased form,
+	// make sure it is lowercased.
+	organization := strings.ToLower(r.flag.Organization)
+
 	appConfig := templateapp.Config{
 		AppName:           appName,
 		Catalog:           r.flag.Catalog,
+		CatalogNamespace:  r.flag.CatalogNamespace,
 		Cluster:           r.flag.Cluster,
 		DefaultingEnabled: r.flag.DefaultingEnabled,
 		InCluster:         r.flag.InCluster,
 		Name:              r.flag.Name,
 		Namespace:         r.flag.Namespace,
+		Organization:      organization,
 		Version:           r.flag.Version,
 	}
 
@@ -69,21 +76,20 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		assetName = key.GenerateAssetName(appName, "userconfig", r.flag.Cluster)
 	}
 
+	// If organization is passed to the command use it as the indicator for
+	// templating App CR in org namespace.
 	var namespace string
 	if r.flag.InCluster {
 		namespace = r.flag.Namespace
+	} else if appConfig.Organization != "" {
+		namespace = fmt.Sprintf("org-%s", appConfig.Organization)
 	} else {
 		namespace = r.flag.Cluster
 	}
 
 	if r.flag.flagUserSecret != "" {
-		userConfigSecretData, err := key.ReadSecretYamlFromFile(afero.NewOsFs(), r.flag.flagUserSecret)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		secretConfig := templateapp.SecretConfig{
-			Data:      userConfigSecretData,
+		secretConfig := templateapp.UserConfig{
+			Path:      r.flag.flagUserSecret,
 			Name:      assetName,
 			Namespace: namespace,
 		}
@@ -100,19 +106,12 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 	}
 
 	if r.flag.flagUserConfigMap != "" {
-		var configMapData string
-		if r.flag.flagUserConfigMap != "" {
-			configMapData, err = key.ReadConfigMapYamlFromFile(afero.NewOsFs(), r.flag.flagUserConfigMap)
-			if err != nil {
-				return microerror.Mask(err)
-			}
-		}
-
-		configMapConfig := templateapp.ConfigMapConfig{
-			Data:      configMapData,
+		configMapConfig := templateapp.UserConfig{
+			Path:      r.flag.flagUserConfigMap,
 			Name:      assetName,
 			Namespace: namespace,
 		}
+
 		userConfigMap, err := templateapp.NewConfigMap(configMapConfig)
 		if err != nil {
 			return microerror.Mask(err)
@@ -142,13 +141,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		return microerror.Mask(err)
 	}
 
-	type AppCROutput struct {
-		AppCR               string
-		UserConfigSecret    string
-		UserConfigConfigMap string
-	}
-
-	appCROutput := AppCROutput{
+	appCROutput := templateapp.AppCROutput{
 		AppCR:               string(appCRYaml),
 		UserConfigConfigMap: string(userConfigConfigMapYaml),
 		UserConfigSecret:    string(userConfigSecretYaml),