+++ cmd/template/cluster/provider/capo.go
@@ -2,15 +2,21 @@ package provider
 
 import (
 	"context"
+	"fmt"
 	"io"
+	"os"
+	"text/template"
 
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
+	"sigs.k8s.io/yaml"
 
 	"github.com/giantswarm/kubectl-gs/cmd/template/cluster/provider/templates/openstack"
+	"github.com/giantswarm/kubectl-gs/internal/key"
+	templateapp "github.com/giantswarm/kubectl-gs/pkg/template/app"
 )
 
-func WriteOpenStackTemplate(ctx context.Context, client k8sclient.Interface, out io.Writer, config ClusterCRsConfig) error {
+func WriteOpenStackTemplateRaw(ctx context.Context, client k8sclient.Interface, out io.Writer, config ClusterCRsConfig) error {
 	data := struct {
 		Description       string
 		KubernetesVersion string
@@ -59,6 +65,100 @@ func WriteOpenStackTemplate(ctx context.Context, client k8sclient.Interface, out
 	}
 
 	err := runMutation(ctx, client, data, templates, out)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+func WriteOpenStackTemplateAppCR(ctx context.Context, config ClusterCRsConfig) error {
+	var userConfigConfigMapYaml []byte
+	var err error
+
+	clusterAppOutput := templateapp.AppCROutput{}
+	defaultAppsAppOutput := templateapp.AppCROutput{}
+
+	clusterAppConfig := templateapp.Config{
+		AppName:   config.Name,
+		Catalog:   config.ClusterAppCatalog,
+		InCluster: true,
+		Name:      "cluster-openstack",
+		Namespace: fmt.Sprintf("org-%s", config.Organization),
+		Version:   config.ClusterAppVersion,
+	}
+
+	defaultAppsAppConfig := templateapp.Config{
+		AppName:   fmt.Sprintf("%s-default-apps", config.Name),
+		Catalog:   config.DefaultAppsAppCatalog,
+		InCluster: true,
+		Name:      "default-apps-openstack",
+		Namespace: fmt.Sprintf("org-%s", config.Organization),
+		Version:   config.DefaultAppsAppVersion,
+	}
+
+	userConfig := templateapp.UserConfig{
+		Namespace: fmt.Sprintf("org-%s", config.Organization),
+	}
+
+	if config.ClusterAppUserConfigMap != "" {
+		userConfig.Name = fmt.Sprintf("%s-userconfig", clusterAppConfig.AppName)
+		userConfig.Path = config.ClusterAppUserConfigMap
+
+		userConfigMap, err := templateapp.NewConfigMap(userConfig)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+		clusterAppConfig.UserConfigConfigMapName = userConfigMap.GetName()
+
+		userConfigConfigMapYaml, err = yaml.Marshal(userConfigMap)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		clusterAppOutput.UserConfigConfigMap = string(userConfigConfigMapYaml)
+	}
+
+	if config.DefaultAppsAppUserConfigMap != "" {
+		userConfig.Name = fmt.Sprintf("%s-userconfig", defaultAppsAppConfig.AppName)
+		userConfig.Path = config.DefaultAppsAppUserConfigMap
+
+		userConfigMap, err := templateapp.NewConfigMap(userConfig)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+		defaultAppsAppConfig.UserConfigConfigMapName = userConfigMap.GetName()
+
+		userConfigConfigMapYaml, err = yaml.Marshal(userConfigMap)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		defaultAppsAppOutput.UserConfigConfigMap = string(userConfigConfigMapYaml)
+	}
+
+	clusterAppCRYaml, err := templateapp.NewAppCR(clusterAppConfig)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	clusterAppOutput.AppCR = string(clusterAppCRYaml)
+
+	defaultAppsAppCRYaml, err := templateapp.NewAppCR(defaultAppsAppConfig)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	defaultAppsAppOutput.AppCR = string(defaultAppsAppCRYaml)
+
+	t := template.Must(template.New("appCR").Parse(key.AppCRTemplate))
+
+	err = t.Execute(os.Stdout, clusterAppOutput)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = t.Execute(os.Stdout, defaultAppsAppOutput)
 	if err != nil {
 		return microerror.Mask(err)
 	}